NODE-1080: Refactor
